Add files via upload
PowerBI basic project that read the name of the stock from "Stock symbols.xlsx" and creates a line graph with both the adjusted closing prices and the  simple moving avarage(SMA) of the last 30 days since the last data colection.

The limitation with this approach is that the spreadsheets containg the stocks info needs to be manually updated. The same project can be done with pandas Data_read easier and more up-to-date way.